cmd: Use genericclioptions
genericclioptions is helper package that makes it easier to develop plugins
that offer a similar experience to kubectl.

This commit uses this package, the main change is that the custom handling of
the namespace is removed everywhere. As a result of this, commands that doesn't
specify a namespace act on the configured namespace in the kubeconfig file.